where_to_go/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-admin-sortable2==0.7.7
django-tinymce==3.0.2
Pillow==7.2.0
python-dotenv==0.9.1
pytz==2020.1
sqlparse==0.3.1
